Melhorias na avaliação

+++ src/gen.cpp
@@ -25,8 +25,8 @@ u64 bit_peao_defende[LADOS][CASAS_DO_TABULEIRO];
 int peao_direita[LADOS][CASAS_DO_TABULEIRO];
 int peao_esquerda[LADOS][CASAS_DO_TABULEIRO];
 
-u64 peao_uma_casa[LADOS][CASAS_DO_TABULEIRO];
-u64 peao_duas_casas[LADOS][CASAS_DO_TABULEIRO];
+int peao_uma_casa[LADOS][CASAS_DO_TABULEIRO];
+int peao_duas_casas[LADOS][CASAS_DO_TABULEIRO];
 
 u64 bit_moves_cavalo[CASAS_DO_TABULEIRO];
 u64 bit_moves_bispo[CASAS_DO_TABULEIRO];
